seguridad de api peluche

+++ RealStateApp.Core.Application/Features/Improvements/Commands/DeleteImprovementById/DeleteImprovementsByIdCommand.cs
@@ -1,15 +1,18 @@
 ﻿
 using MediatR;
+using RealStateApp.Core.Application.Exceptions;
 using RealStateApp.Core.Application.Interface.Repositories;
+using RealStateApp.Core.Application.Wrappers;
+using System.Net;
 
 namespace RealStateApp.Core.Application.Features.Improvements.Commands.DeleteImprovementById
 {
-    public class DeleteImprovementsByIdCommand : IRequest<int>
+    public class DeleteImprovementsByIdCommand : IRequest<Response<int>>
     {
         public int Id { get; set; } 
     }
 
-    public class DeleteImprovementsByIdCommandHandler : IRequestHandler<DeleteImprovementsByIdCommand, int>
+    public class DeleteImprovementsByIdCommandHandler : IRequestHandler<DeleteImprovementsByIdCommand, Response<int>>
     {
         private readonly IImprovementsRepository _repository;
 
@@ -18,13 +21,15 @@ namespace RealStateApp.Core.Application.Features.Improvements.Commands.DeleteImp
             _repository = repository;
         }
 
-        public async Task<int> Handle(DeleteImprovementsByIdCommand command, CancellationToken cancellationToken)
+        public async Task<Response<int>> Handle(DeleteImprovementsByIdCommand command, CancellationToken cancellationToken)
         {
             var get = await _repository.GetByIdAsync(command.Id);
 
+            if (get == null) throw new ApiException("Improvements not found", (int)HttpStatusCode.NotFound);
+
             await _repository.DeleteAsync(get);
 
-            return get.Id;
+            return new Response<int>(get.Id);
         }
